Move core classes by namespaces
