+++ AsyncImageLoader.Avalonia.Demo/Services/LongLoader.cs
@@ -1,6 +1,8 @@
 ﻿using System.Threading;
 using System.Threading.Tasks;
 using AsyncImageLoader.Core;
+using AsyncImageLoader.Core.Leases;
+using AsyncImageLoader.Core.Pipeline;
 using AsyncImageLoader.Loaders;
 
 namespace AsyncImageLoader.Avalonia.Demo.Services;
