+++ AsyncImageLoader.Avalonia.Tests/BaseWebImageLoaderTests.cs
@@ -5,6 +5,7 @@ using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
 using AsyncImageLoader.Core;
+using AsyncImageLoader.Core.Pipeline;
 using AsyncImageLoader.Loaders;
 using Avalonia.Media.Imaging;
 using AwesomeAssertions;
