+++ AsyncImageLoader.Avalonia.Tests/BitmapDecoderTests.cs
@@ -3,6 +3,7 @@ using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
 using AsyncImageLoader.Core;
+using AsyncImageLoader.Core.Decoding;
 using AwesomeAssertions;
 using Xunit;
