+++ AsyncImageLoader.Avalonia.Tests/DiskCachedWebImageLoaderTests.cs
@@ -4,6 +4,7 @@ using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
 using AsyncImageLoader.Core;
+using AsyncImageLoader.Core.Pipeline;
 using AsyncImageLoader.Loaders;
 using AwesomeAssertions;
 using Xunit;
