+++ AsyncImageLoader.Avalonia.Tests/DiskImageByteCacheTests.cs
@@ -3,6 +3,7 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using AsyncImageLoader.Core;
+using AsyncImageLoader.Core.Caching;
 using AwesomeAssertions;
 using Xunit;
