+++ AsyncImageLoader.Avalonia.Tests/HttpImageTransportTests.cs
@@ -5,6 +5,8 @@ using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
 using AsyncImageLoader.Core;
+using AsyncImageLoader.Core.Pipeline;
+using AsyncImageLoader.Core.Transport;
 using AwesomeAssertions;
 using Xunit;
