+++ AsyncImageLoader.Avalonia.Tests/ImageLoaderPipelineBuilderTests.cs
@@ -4,6 +4,10 @@ using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
 using AsyncImageLoader.Core;
+using AsyncImageLoader.Core.Decoding;
+using AsyncImageLoader.Core.Pipeline;
+using AsyncImageLoader.Core.Sources;
+using AsyncImageLoader.Core.Transport;
 using Avalonia.Media.Imaging;
 using AwesomeAssertions;
 using Xunit;
