+++ AsyncImageLoader.Avalonia.Tests/ImageRequestCoordinatorTests.cs
@@ -1,5 +1,7 @@
 using System;
 using AsyncImageLoader.Core;
+using AsyncImageLoader.Core.Leases;
+using AsyncImageLoader.Core.Pipeline;
 using Avalonia;
 using Avalonia.Media;
 using AwesomeAssertions;
