+++ AsyncImageLoader.Avalonia/Core/Caching/IImageByteCache.cs
@@ -2,7 +2,7 @@ using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace AsyncImageLoader.Core;
+namespace AsyncImageLoader.Core.Caching;
 
 /// <summary>
 /// Stores encoded image data by cache key.
