+++ AsyncImageLoader.Avalonia/Core/Caching/TransientImageCache.cs
@@ -1,9 +1,10 @@
 using System;
 using System.Threading;
 using System.Threading.Tasks;
+using AsyncImageLoader.Core.Leases;
 using Avalonia.Media;
 
-namespace AsyncImageLoader.Core;
+namespace AsyncImageLoader.Core.Caching;
 
 /// <summary>
 /// Decodes a new image for every request without retaining it in a cache.
