+++ AsyncImageLoader.Avalonia/IAsyncImageLoader.cs
@@ -2,6 +2,8 @@
 using System.Threading;
 using System.Threading.Tasks;
 using AsyncImageLoader.Core;
+using AsyncImageLoader.Core.Leases;
+using AsyncImageLoader.Core.Pipeline;
 
 namespace AsyncImageLoader;
