+++ AsyncImageLoader.Avalonia/ImageLoader.cs
@@ -3,6 +3,8 @@ using System.Runtime.CompilerServices;
 using System.Threading;
 using System.Threading.Tasks;
 using AsyncImageLoader.Core;
+using AsyncImageLoader.Core.Leases;
+using AsyncImageLoader.Core.Pipeline;
 using AsyncImageLoader.Loaders;
 using Avalonia;
 using Avalonia.Controls;
