+++ AsyncImageLoader.Avalonia/Loaders/BaseWebImageLoader.cs
@@ -3,6 +3,8 @@ using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
 using AsyncImageLoader.Core;
+using AsyncImageLoader.Core.Leases;
+using AsyncImageLoader.Core.Pipeline;
 
 namespace AsyncImageLoader.Loaders;
