+++ AsyncImageLoader.Avalonia/Loaders/DiskCachedWebImageLoader.cs
@@ -3,6 +3,9 @@ using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
 using AsyncImageLoader.Core;
+using AsyncImageLoader.Core.Caching;
+using AsyncImageLoader.Core.Leases;
+using AsyncImageLoader.Core.Pipeline;
 
 namespace AsyncImageLoader.Loaders;
